inline QueryRequester implementation
